// frontend-mobile/src/features/iam/clients/data/client.types.ts
import { BaseEntity } from '@/utils';

export interface Client extends BaseEntity {
  name: string;
  subsidiaryId?: string | { _id: string; name: string; recordId: string };
  isOptimistic?: boolean;
}

export interface CreateClientDto {
  _id?: string;
  name: string;
  subsidiaryId?: string;
}

export interface UpdateClientDto {
  name?: string;
  isActive?: boolean;
  __v: number;
}

export interface ClientQuery {
  name?: string;
  includeInactives?: boolean;
}